requirements.txt
pysha3
requests
six
monero-python

